+++ CHANGELOG.md
@@ -1,5 +1,21 @@
 # Changelog
 
+## v1.1.0 (2019-02-09)
+
+- Allow using hooks inside React component decorators, such as `React.memo` or `React.forwardRef`.
+
+  For example, the following code sample now **does not** violate the rule:
+
+  ```tsx
+  const MyComponent = React.memo(props => {
+    useEffect(() => {
+      console.log('Counter changed');
+    }, [props.value]);
+
+    return <div>Counter: {props.value}</div>;
+  });
+  ```
+
 ## v1.0.1 (2019-02-03)
 
 - Updated README